style: message 도메인 일관되게 코드 포맷팅
