+++ src/main/java/com/project/coalba/domain/message/controller/StaffMessageController.java
@@ -4,17 +4,20 @@ import com.project.coalba.domain.message.dto.request.MessageRequest;
 import com.project.coalba.domain.message.dto.response.MessageResponse;
 import com.project.coalba.domain.message.service.StaffMessageService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-@RequestMapping("/staff/messages")
 @RequiredArgsConstructor
+@RequestMapping("/staff/messages")
 @RestController
 public class StaffMessageController {
     private final StaffMessageService staffMessageService;
 
     @PostMapping("/workspaces/{workspaceId}")
-    public void sendMessageToBoss(@PathVariable Long workspaceId, @RequestBody MessageRequest request){
+    public ResponseEntity<Void> sendMessageToBoss(@PathVariable Long workspaceId, @RequestBody MessageRequest request){
         staffMessageService.sendMessageToBoss(workspaceId, request.getContent());
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @GetMapping("/workspaces/{workspaceId}")
