--- src/main/java/com/project/coalba/domain/message/repository/MessageRepositoryCustom.java
@@ -1,10 +1,0 @@
-package com.project.coalba.domain.message.repository;
-
-import com.project.coalba.domain.message.entity.Message;
-
-import java.util.List;
-
-public interface MessageRepositoryCustom {
-
-    List<Message> getMessages(Long staffId, Long workspaceId);
-}
