+++ src/main/java/com/project/coalba/domain/message/repository/custom/MessageRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.project.coalba.domain.message.repository.custom;
+
+import com.project.coalba.domain.message.entity.Message;
+
+import java.util.List;
+
+public interface MessageRepositoryCustom {
+
+    List<Message> getMessages(Long workspaceId, Long staffId);
+}
