+++ src/main/java/com/project/coalba/domain/message/service/StaffMessageService.java
@@ -6,7 +6,6 @@ import com.project.coalba.domain.message.entity.enums.Criteria;
 import com.project.coalba.domain.message.repository.MessageRepository;
 import com.project.coalba.domain.profile.entity.Staff;
 import com.project.coalba.domain.workspace.entity.Workspace;
-import com.project.coalba.domain.workspace.repository.WorkspaceRepository;
 import com.project.coalba.domain.workspace.service.BossWorkspaceService;
 import com.project.coalba.global.utils.ProfileUtil;
 import lombok.RequiredArgsConstructor;
@@ -18,15 +17,14 @@ import java.util.List;
 @RequiredArgsConstructor
 @Service
 public class StaffMessageService {
-    private final MessageRepository messageRepository;
     private final BossWorkspaceService bossWorkspaceService;
+    private final MessageRepository messageRepository;
     private final ProfileUtil profileUtil;
 
     @Transactional
     public void sendMessageToBoss(Long workspaceId, String content) {
         Staff staff = profileUtil.getCurrentStaff();
         Workspace workspace = bossWorkspaceService.getWorkspace(workspaceId);
-
         Message message = Message.builder()
                 .content(content)
                 .criteria(Criteria.STAFF_TO_WORKSPACE)
@@ -41,8 +39,7 @@ public class StaffMessageService {
     public MessageResponse.StaffMessageResponse getDetailMessages(Long workspaceId){
         Long staffId = profileUtil.getCurrentStaff().getId();
         Workspace workspace = bossWorkspaceService.getWorkspace(workspaceId);
-
-        List<Message> messages = messageRepository.getMessages(staffId, workspaceId);
+        List<Message> messages = messageRepository.getMessages(workspaceId, staffId);
         return new MessageResponse.StaffMessageResponse(workspace, messages);
     }
 }
